unvalued attribute test
