+++ package.json
@@ -63,6 +63,6 @@
     "toposort": "^1.0.0"
   },
   "peerDependencies": {
-    "webpack": "1 || ^2.1.0-beta"
+    "webpack": "1 || ^2 || ^2.1.0-beta || ^2.2.0-rc"
   }
 }
